requirements.txt
google-cloud-storage==2.10.0
requests==2.31.0
functions-framework==3.4.0